% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_source_file.R
\name{doc_source_file}
\alias{doc_source_file}
\title{Source a file and document tests performed}
\usage{
doc_source_file(path, env = test_env(), chdir = TRUE,
  encoding = "UTF-8", wrap = TRUE)
}
\arguments{
\item{path}{Path to tests}

\item{env}{Environment in which to evaluate code.}

\item{chdir}{Change working directory to `dirname(path)`?}

\item{encoding}{Deprecated}

\item{wrap}{Automatically wrap all code within [test_that()]? This ensures
that all expectations are reported, even if outside a test block.}
}
\description{
Source a file and document tests performed
}
\keyword{internal}
