% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_test_dir.R
\name{doc_test_dir}
\alias{doc_test_dir}
\title{Run all tests in directory or package and document results}
\usage{
doc_test_dir(path, filter = NULL, env = test_env(), ...,
  encoding = "unknown", load_helpers = TRUE, stop_on_failure = FALSE,
  stop_on_warning = FALSE, wrap = TRUE)
}
\arguments{
\item{path}{path to tests}

\item{filter}{If not `NULL`, only tests with file names matching this
regular expression will be executed.  Matching will take on the file
name after it has been stripped of `"test-"` and `".R"`.}

\item{...}{Additional arguments passed to [grepl()] to control filtering.
warnings.}

\item{package}{package name}
}
\description{
Use `doc_test_dir()` for a collection of tests in a directory

Provides an alternate to the test_dir functionality in the `testhat` package,
and generates more thorough documentation surrounding the tests.
}
\examples{
\dontrun{test_package("testDoctor")}
}
