% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_test_file.R
\name{doc_test_file}
\alias{doc_test_file}
\title{Run all tests in specified file.}
\usage{
doc_test_file(path, env = test_env(), load_helpers = TRUE,
  encoding = "UTF-8", wrap = TRUE)
}
\arguments{
\item{path}{path to file}

\item{env}{environment in which to execute the tests}

\item{load_helpers}{Source helper files before running the tests?}

\item{encoding}{File encoding, default is "unknown"
`unknown`.}

\item{wrap}{Automatically wrap all code within [test_that()]? This ensures
that all expectations are reported, even if outside a test block.}

\item{stop_on_failure}{If `TRUE`, throw an error if any tests fail.}

\item{stop_on_warning}{If `TRUE`, throw an error if any tests generate}
}
\value{
the results as a "testthat_results" (list)
}
\description{
Run all tests in specified file.
}
