% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_test_files.R
\name{doc_test_files}
\alias{doc_test_files}
\title{Run all tests listed in the files}
\usage{
doc_test_files(paths, env = test_env(), stop_on_failure = FALSE,
  stop_on_warning = FALSE, wrap = TRUE)
}
\arguments{
\item{paths}{paths to each test file}

\item{env}{testing env
warnings.}

\item{stop_on_failure}{If `TRUE`, throw an error if any tests fail.}

\item{stop_on_warning}{If `TRUE`, throw an error if any tests generate}

\item{wrap}{Automatically wrap all code within [test_that()]? This ensures
that all expectations are reported, even if outside a test block.}
}
\description{
Evaluates each of the tests in each file passed to paths
}
\examples{
\dontrun{doc_test_dir("tests/testthat")}
}
