% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doctor_checkup.R
\name{doctor_checkup}
\alias{doctor_checkup}
\title{Run all tests in directory or package, and generage nice documentation}
\usage{
doctor_checkup(pkg = ".", document = "", compile = NULL)
}
\arguments{
\item{pkg}{path to package}

\item{document}{the name of the document to write to. If left as "", writes to console.}

\item{compile}{the name of the document to write to. If left as "", writes to console.}
}
\description{
Use `doctor_checkup()` for a collection of tests in a directory;

Framework is built to work on top of the `testthat` library, so no changes to your tests need to be made to make this work.
Your outputs will be saved to a markdown document in the tests folder
}
