% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doctored_test_results.R
\name{doctored_test_results}
\alias{doctored_test_results}
\title{print the results of the Tests}
\usage{
doctored_test_results(desc, code, scribe)
}
\arguments{
\item{desc}{description of test, from test_that}

\item{code}{test code to be converted to text from test_that}

\item{scribe}{test_scribe object containing results of test}
}
\description{
print the results of the Tests
}
