% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_scribe.R
\docType{data}
\name{test_scribe}
\alias{test_scribe}
\title{object for managing the recording of tests and their results
'}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
test_scribe
}
\description{
object for managing the recording of tests and their results
'
}
\keyword{internal}
